requirements.txt
babel==2.17.0
PyYAML==6.0.2
